agentic_RAG_with_langgraph/requirements.txt
python-dotenv==1.1.0
flask==3.1.0
openai==1.76.2
pydantic==2.11.4
langgraph==0.4.5
chromadb==1.0.10